rename pi-foundation to raspberrypi to be consistent with U-Boot and Linux naming scheme
