+++ board/raspberrypi/pizero/pizero.go
@@ -17,7 +17,7 @@ package pizero
 import (
 	_ "unsafe"
 
-	"github.com/f-secure-foundry/tamago/board/pi-foundation"
+	"github.com/f-secure-foundry/tamago/board/raspberrypi"
 	"github.com/f-secure-foundry/tamago/soc/bcm2835"
 )
